Made changes to optimized the count query for the ado.net context
